Add support for save/load of enumerated types
By default, we now try to cast to/from int when handling types
with now explicit save_override/load_override. This gives us
automatic support for enumerated types, which can always be cast
to/from int.

If casting to/from int fails, we fall back to previous handling which
just reports that the type in question is not supported.

Also, we now throw exceptions rather than calling abort() on errors.